requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
PyJWT
requests